+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/ClouderaManagerRoleRefreshService.java
@@ -1,6 +1,5 @@
 package com.sequenceiq.cloudbreak.cm;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
@@ -65,7 +64,7 @@ class ClouderaManagerRoleRefreshService {
         LOGGER.debug("Wait if Generate Credentials command is still active.");
         ClouderaManagerResourceApi clouderaManagerResourceApi = clouderaManagerApiFactory.getClouderaManagerResourceApi(client);
         ApiCommandList apiCommandList = clouderaManagerResourceApi.listActiveCommands(DataView.SUMMARY.name());
-        Optional<BigDecimal> generateCredentialsCommandId = apiCommandList.getItems().stream()
+        Optional<Long> generateCredentialsCommandId = apiCommandList.getItems().stream()
                 .filter(toGenerateCredentialsCommand()).map(ApiCommand::getId).findFirst();
         generateCredentialsCommandId.ifPresent(pollCredentialGeneration(stack, client));
     }
@@ -74,7 +73,7 @@ class ClouderaManagerRoleRefreshService {
         return apiCommand -> GENERATE_CREDENTIALS_COMMAND_NAME.equals(apiCommand.getName());
     }
 
-    private Consumer<BigDecimal> pollCredentialGeneration(StackDtoDelegate stack, ApiClient client) {
+    private Consumer<Long> pollCredentialGeneration(StackDtoDelegate stack, ApiClient client) {
         return id -> {
             LOGGER.debug("Generate Credentials command is still active.");
             clouderaManagerPollingServiceProvider.startPollingCmGenerateCredentials(stack, client, id);
