+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/ClouderaManagerSecurityService.java
@@ -1,6 +1,5 @@
 package com.sequenceiq.cloudbreak.cm;
 
-import java.math.BigDecimal;
 import java.nio.charset.Charset;
 import java.security.KeyPair;
 import java.util.List;
@@ -473,7 +472,7 @@ public class ClouderaManagerSecurityService implements ClusterSecurityService {
 
     private void processHostCertsBatchResponse(ApiClient client, ApiBatchResponse apiBatchResponse) {
         if (apiBatchResponse != null && apiBatchResponse.isSuccess() != null && apiBatchResponse.getItems() != null && apiBatchResponse.isSuccess()) {
-            List<BigDecimal> ids = apiBatchResponse.getItems().stream()
+            List<Long> ids = apiBatchResponse.getItems().stream()
                     .map(bre -> new Json((String) bre.getResponse()).getUnchecked(ApiCommand.class).getId())
                     .collect(Collectors.toList());
             ExtendedPollingResult pollingResult = clouderaManagerPollingServiceProvider.startPollingCommandList(stack, client, ids, "Rotate host certificates");
@@ -490,5 +489,4 @@ public class ClouderaManagerSecurityService implements ClusterSecurityService {
     private String createPrivateKeyString(KeyPair sshKeyPair) {
         return StringUtils.removeEnd(PkiUtil.convert(sshKeyPair.getPrivate()), "\n");
     }
-
 }
