+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/polling/task/AbstractClouderaManagerCommandListCheckerTask.java
@@ -1,6 +1,5 @@
 package com.sequenceiq.cloudbreak.cm.polling.task;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -44,7 +43,7 @@ public abstract class AbstractClouderaManagerCommandListCheckerTask<T extends Cl
 
     private List<ApiCommand> collectApiCommands(T pollerObject, CommandsResourceApi commandsResourceApi) throws ApiException {
         List<ApiCommand> apiCommands = new ArrayList<>();
-        for (BigDecimal commandId : pollerObject.getIdList()) {
+        for (Long commandId : pollerObject.getIdList()) {
             ApiCommand apiCommand = commandsResourceApi.readCommand(commandId);
             apiCommands.add(apiCommand);
             if (apiCommand.isActive()) {
